requirements.txt
langchain==0.2.17
langchain-community==0.2.19
langchain-experimental==0.0.65

langchain-groq==0.1.9
langchain-pinecone==0.1.3

pinecone[grpc]>=5.1,<6

sentence-transformers>=2.6,<3

flask>=3.0,<4
pypdf>=4.0,<5
python-dotenv>=1.0,<2

-e .